requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2
django-autoslug==1.9.8
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-js-asset==1.2.2
Pillow==9.5.0
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
